occupancy_networks/scripts/dataset_mito/simulation/sim_env_requirement.txt
absl-py==1.1.0
alabaster==0.7.12
anyio==2.2.0
appdirs==1.4.4
apptools==5.1.0
argon2-cffi==20.1.0
asciitree==0.3.3
asttokens==2.0.8
async-generator==1.10
attrs==20.3.0
Automat==0.8.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
blessings==1.7
blinker==1.4
cachetools==5.2.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
cloud-init==22.3.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.17.1
entrypoints==0.3
envisage==5.0.0
et-xmlfile==1.1.0
executing==1.0.0
fasteners==0.18
filelock==3.0.12
freetype-py==2.2.0
fsspec==2022.10.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grpcio==1.46.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imagecodecs==2022.9.26
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==4.11.4
importlib-resources==5.1.2
incremental==16.10.1
iotop==0.6
ipdb==0.13.7
ipykernel==5.5.3
ipympl==0.7.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.2
jupyter-server==1.6.4
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.0
jupyterlab-widgets==1.0.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.37.0
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==1.1.1
matlabengineforpython===R2021a
matplotlib==3.4.1
matplotlib-inline==0.1.2
mayavi==4.7.2
mistune==0.8.4
mlnx-tools===-5.2.0-
more-itertools==4.2.0
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
notebook==6.3.0
numba==0.54.0
numcodecs==0.10.2
numpy==1.20.2
numpydoc==1.1.0
nvidia-ml-py3==7.352.0
nvsm==22.3.7
nvsm-dshm==22.3.7
nvsm-health==22.3.7
nvsm-sdk==1.0.0
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.8
packaging==20.9
paho-mqtt==1.5.0
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.19.4
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pyface==7.3.0
Pygments==2.9.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scikit-image==0.18.1
scipy==1.6.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
skan==0.10.0.dev0
sniffio==1.2.0
snowballstemmer==2.1.0
sos==4.4
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.10
stack-data==0.5.0
systemd-python==234
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.9.4
testpath==0.4.4
tifffile==2021.4.8
toml==0.10.2
tomlkit==0.7.0
toolz==0.11.1
torch-tb-profiler==0.4.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
traits==6.2.0
traitsui==7.1.1
Twisted==18.9.0
typing-extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
virtualenv==20.4.4
vispy==0.6.6
vtk==9.0.1
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.5.1
xkit==0.0.0
zarr==2.13.3
zipp==3.4.1
zope.interface==4.7.1
